Trajectory tracking working, but there's a weird bug about the position moving a little bit once it's at the final position

+++ positioncontrol.h
@@ -7,7 +7,7 @@ int positionCtrlInit();
 
 float setPositionGains(float kp, float kd, float ki);
 PIDInfo getPositionGains();
-int addTrajPoint(int point, int index);
+float addTrajPoint(float point, int index);
 void setHoldAngle(float angle);
